Fix ambiguous-template-usage in Elf.h

+++ src/Elf.h
@@ -259,7 +259,7 @@ public:
         this->size = size;
         return *this;
     }
-    uint64_t get_size() const { return std::max(size, contents.size()); }
+    uint64_t get_size() const { return std::max((uint64_t) size, (uint64_t) contents.size()); }
     ///@}
 
     Section &set_alignment(uint64_t alignment) {
